Rename recommender to recommendation package name
