+++ src/main/java/seedu/address/model/EduMate.java
@@ -10,8 +10,8 @@ import javafx.collections.ObservableList;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.UniquePersonList;
 import seedu.address.model.person.User;
-import seedu.address.model.recommender.Recommendation;
-import seedu.address.model.recommender.UniqueRecommendationList;
+import seedu.address.model.recommendation.Recommendation;
+import seedu.address.model.recommendation.UniqueRecommendationList;
 import seedu.address.model.util.SampleDataUtil;
 
 /**
