+++ src/main/java/seedu/address/model/recommendation/exceptions/DuplicateRecommendationException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.recommender.exceptions;
+package seedu.address.model.recommendation.exceptions;
 
 /**
  * Signals that the operation will result in duplicate Persons (Persons are considered duplicates if they have the same
