+++ src/main/java/seedu/address/model/recommendation/exceptions/RecommendationNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.recommender.exceptions;
+package seedu.address.model.recommendation.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified person.
