+++ src/main/java/seedu/address/storage/JsonAdaptedRecommendation.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.location.Location;
-import seedu.address.model.recommender.Recommendation;
+import seedu.address.model.recommendation.Recommendation;
 import seedu.address.model.tag.ModuleTag;
 import seedu.address.model.time.TimePeriod;
